refactor: remove container from headerbar
